Grid search result picker works
